clean up kanban directory move scss files to styles directory componentize components into their own files delete unused vars and exports
