+++ src/components/Kanban/TicketDialog/Buttons/index.ts
@@ -1,3 +1,4 @@
+export * from './CancelTicketBtn'
 export * from './CreateTicketBtn'
 export * from './DeleteTicketBtn'
 export * from './SaveTicketBtn'
