+++ src/components/Kanban/TicketDialog/Dialogs/index.ts
@@ -1,0 +1,2 @@
+export * from './CancelDialog'
+export * from './DeleteDialog'
