+++ src/components/Kanban/TicketDialog/Fields/index.ts
@@ -1,4 +1,6 @@
 export * from './SelectDate'
 export * from './SelectStatus'
 export * from './SelectAssignee'
+export * from './TicketCreator'
+export * from './TicketDropdownFields'
 export * from './TicketTextFields'
