+++ src/components/Kanban/TicketDialog/TicketDialog.tsx
@@ -1,36 +1,18 @@
 import { Dialog, DialogContent } from '@mui/material'
 import { useAppDispatch, useAppSelector } from 'utils/redux/hooks'
+import { TicketDropdownFields, TicketTextFields } from './Fields'
 import {
-  SelectAssignee,
-  SelectDate,
-  SelectStatus,
-  TicketTextFields,
-  TicketTextLabel,
-} from './Fields'
-import {
-  selectConfirmationDialogType,
   selectTicketDialogState,
   selectTicketFields,
   selectVisibleTicketDialog,
 } from 'utils/redux/slices/taskBoardSlice'
-import {
-  closeConfirmationDialog,
-  closeVisibleTicketDialog,
-  toggleCancelDialog,
-} from 'utils/helpers/taskHelpers'
-import { TicketDialogButtons } from './Buttons'
-import { Comments } from './Comments/Comments'
-import './TicketDialog.scss'
-import '../styles/ConfirmationDialogs.scss'
-import { TeamAvatar } from 'components/TeamAvatar/TeamAvatar'
-import {
-  selectMembersById,
-  selectProjectId,
-} from 'utils/redux/slices/projectSlice'
-import { useEffect, useState } from 'react'
-import { PrimaryButton, SecondaryButton } from 'components/Buttons'
+import { toggleCancelDialog } from 'utils/helpers/taskHelpers'
+import { Comments } from './Comments'
+import { selectProjectId } from 'utils/redux/slices/projectSlice'
 import { fetchIcon } from 'utils/components/Icons'
 import { selectUserId } from 'utils/redux/slices/userSlice'
+import { CancelDialog, DeleteTicketDialog } from './Dialogs'
+import '../styles/TicketDialog.scss'
 
 export const TicketDialog = () => {
   const projectId = useAppSelector(selectProjectId)
@@ -62,81 +44,7 @@ export const TicketDialog = () => {
         </DialogContent>
       </Dialog>
       <CancelDialog />
+      <DeleteTicketDialog />
     </>
-  )
-}
-
-const CancelDialog = () => {
-  const confirmationDialogType = useAppSelector(selectConfirmationDialogType)
-  const dispatch = useAppDispatch()
-  const handleCloseVisibleTicketDialog = () =>
-    closeVisibleTicketDialog(dispatch)
-  const handleCloseDialog = () => closeConfirmationDialog(dispatch)
-  const openDialog = confirmationDialogType === 'cancel'
-
-  return (
-    <Dialog open={openDialog} onClose={handleCloseDialog} maxWidth='xs'>
-      <DialogContent className='confirmation-dialog'>
-        <h3>Close this ticket?</h3>
-        <p>Any information you input or changes you made will not be saved.</p>
-        <div className='buttons'>
-          <SecondaryButton
-            handler={handleCloseDialog}
-            text='Cancel'
-            variant='text'
-          />
-          <PrimaryButton
-            disableElevation
-            handler={handleCloseVisibleTicketDialog}
-            text='Close'
-          />
-        </div>
-      </DialogContent>
-    </Dialog>
-  )
-}
-
-// TODO: Work on combining styles between this and ticket text
-const TicketDropdownFields = () => {
-  return (
-    <div className='ticket-dropdown-fields'>
-      <SelectStatus />
-      <TicketCreator />
-      <SelectAssignee />
-      <SelectDate />
-      <TicketDialogButtons />
-    </div>
-  )
-}
-
-const TicketCreator = () => {
-  const { createdBy } = useAppSelector(selectTicketFields)
-  const [user] = useAppSelector(selectMembersById([createdBy]))
-  const [creator, setCreator] = useState({
-    userId: 'Unassigned',
-    firstName: '',
-    lastName: '',
-    role: '',
-  })
-  const { firstName, lastName, role, userId } = creator
-
-  useEffect(() => {
-    if (user) {
-      const { firstName, lastName, role, _id: userId } = user
-      setCreator({ firstName, lastName, role, userId })
-    }
-  }, [])
-
-  return (
-    <div>
-      <TicketTextLabel icon='person' label='Created by' />
-      <div className='ticket-creator'>
-        <TeamAvatar userId={userId} size='small' />
-        <div className='creator-info'>
-          <p className='name'>{`${firstName} ${lastName}`}</p>
-          <p className='role'>{role}</p>
-        </div>
-      </div>
-    </div>
   )
 }
