+++ src/components/Kanban/index.ts
@@ -1,4 +1,1 @@
-export * from './TaskBoard/CreateTicketTab'
-export * from './TicketDialog/Fields'
-export * from './TicketDialog/TicketDialog'
 export * from './TaskBoard/TaskBoard'
