+++ src/components/Kanban/styles/Comments.scss
