+++ src/components/Kanban/styles/TicketDialog.scss
